dispatcher/backend/requirements.txt
celery==4.1.0
PyJWT==1.5.3
pymongo==3.5.1